Revert to original vercel deployment pattern
