+++ .gitignore
@@ -2,4 +2,4 @@ node_modules
 .env
 .yarn
 .vercel
-dist
+# dist
