+++ src/resolvers/ImageResolver.js
