+++ src/resolvers/SeriesResolver.js
