+++ src/schemas/tvdb.graphql
@@ -1,12 +1,3 @@
-import gql from 'graphql-tag';
-import { readFileSync } from 'fs';
-import { fileURLToPath } from 'url';
-import { dirname, resolve, join } from 'path';
-
-const modulePath = dirname(fileURLToPath(import.meta.url))
-// const typeDefinitions = readFileSync(resolve(modulePath, '../schemas/tvdb.graphql'), 'utf8');
-
-export default gql`
 input Auth {
   apikey: String!
 }
@@ -408,5 +399,3 @@ type Query {
 type Mutation {
   login(auth: Auth!): Token
 }
-
-`;
